refactor: centralize models package
