+++ services/api/app/db.py
@@ -5,7 +5,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
 from .config import get_settings
-from .models import Base
+from services.common.models import Base
 
 
 _engine = None
