+++ services/api/tests/test_outliers.py
@@ -12,7 +12,7 @@ os.environ["DATABASE_URL"] = "sqlite:///./test.db"
 
 from services.api.app.main import app  # noqa: E402
 from services.api.app.db import SessionLocal, engine, get_db  # noqa: E402
-from services.api.app.models import Base, Track, Artist, Listen, MoodScore  # noqa: E402
+from services.common.models import Base, Track, Artist, Listen, MoodScore  # noqa: E402
 from services.api.app.constants import AXES, DEFAULT_METHOD  # noqa: E402
 
 Base.metadata.create_all(bind=engine)
